+++ src/Routes.js
@@ -1,22 +1,13 @@
-import React from "react";
+// @flow
+import * as React from "react";
 import { BrowserRouter, Route } from "react-router-dom";
-
-import Singleton from './data/Singleton';
 
 import Entrar from "./pages/entrar/Entrar";
 import Home from "./pages/home/Home";
 
-class Routes extends React.Component {
+class Routes extends React.Component<{}> {
 
-    constructor(props) {
-        super(props);
-
-        this.state = {
-            jogador: String(Singleton.jogador.nome).length > 3,
-        };
-    }
-
-    render() {
+    render(): React.Node {
         return (
             <BrowserRouter>
                 <Route exact path="/" component={Home} />