+++ src/service/sala.js
@@ -1,0 +1,22 @@
+// @flow
+// import store from '../store/index';
+
+/**
+ * Cria uma nova sala de jogo. Ao criar a sala, atualiza o código da sala no controlador de
+ * estados e retorna um booleano de valor verdadeiro.
+ * 
+ * @returns boolean
+ */
+export const criarSala = (): boolean => {
+    return false;
+};
+
+/**
+ * Busca uma sala de jogo, retornando um booleado com o valor verdadeiro caso a sala seja
+ * encontrada.
+ * 
+ * @returns boolean
+ */
+export const buscarSala = (codigo: string): boolean => {
+    return false;
+}