+++ src/store/reducers/index.js
@@ -1,0 +1,9 @@
+import { combineReducers } from 'redux';
+
+import jogadorReducer from './jogadorReducer';
+import salaReducer from './salaReducer';
+
+export default combineReducers({
+    jogadorReducer,
+    salaReducer
+});