+++ CHANGELOG.md
@@ -6,8 +6,9 @@ Seu formato é baseado no [Keep a Changelog](https://keepachangelog.com/en/1.0.0
 ## [Não publicado]
 ### Adicionado
 - Projeto criado com base no [Create React App](https://github.com/facebook/create-react-app).
-- Controle de dados via Singleton adicionado.
-- Gerenciamento de rotas adicionado.
+- Controle de dados via [Redux](https://react-redux.js.org/) adicionado.
+- Gerenciamento de rotas via [ReactRouter](https://reactrouter.com/web) adicionado.
+- Notificações via [React Toast Notifications](https://jossmac.github.io/react-toast-notifications/) adicionadas.
 - Tipagem via [Flow](https://flow.org/) adicionada.
 
 [Não publicado]: https://github.com/mateusbmp/blackjack-frontend